+++ packages/flutter_hooks/CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased patch
+- Added `useOnListenableChange` (thanks to @whynotmake-it)
+
 ## 0.21.1-pre.2 - 2024-07-22
 
 - Added `onAttach` and `onDetach` to `useScrollController` and `usePageController` (thanks to @whynotmake-it)